# IRowInputFormat performance fix