@oandrew 2020-04-26
@alexey-milovidov  is it even possible it affects performance that much?