@alexey-milovidov 2020-04-26
@oandrew Yes, creating temporary std::vector/std::string/whatever in a loop is a common performance antipattern.